+++ SandboxiePlus/SandMan/Wizards/SetupWizard.h
@@ -123,6 +123,7 @@ private:
     QCheckBox *m_pContecxtMenu;
     QCheckBox *m_pBrowserIcon;
 	QCheckBox* m_pEditOnlyAdmin;
+	QCheckBox* m_pImproveGame;
 };
 
 //////////////////////////////////////////////////////////////////////////////////////////